refactor(Product): id, price int -> long 타입으로 변경

+++ src/main/java/gift/Product.java
@@ -1,5 +1,5 @@
 package gift;
 
-public record Product(int id, String name, int price, String imageUrl) {
+public record Product(long id, String name, long price, String imageUrl) {
 
 }
